Made a new method to export file to an external device. When a lot of records needs to be exported into a device the application takes a lot of time to do it. To avoid this issue a new exportation method was created. Now it works like this: A Qtimer is started within ExportLocalData class constructor. When this timer emmit a timeout sinal is called a function that is responsible for look into database and search for not synced data and write this data into a temporary file. When a device is connected is called a function that will copy this file into external device. This method is faster and saves some time.
Another change was the verification of path where the device was mounted. In earlier versions the application reads all directories in /media and tries to export in all of them, until done.
Now when a device is detected the application reads the /proc/mouts file (shows information about all file systems mounted on kernel). Using a RegEx the application know all devices mounted in /media directory.
Using the mount point verifies if this path is writable. If it is export data into it, otherwise create an error log record.

Besides that, now the application verifies if the modules directory exists. It looks into current application directory and one directory up.
If it doesn't found then the application crashes.